move transform_sort_merge to common_pipeline
