+++ src/common/pipeline/transforms/src/processors/transforms/mod.rs
@@ -17,6 +17,7 @@ pub mod transform_block_compact;
 pub mod transform_compact;
 pub mod transform_expression;
 pub mod transform_expression_executor;
+pub mod transform_sort_merge;
 pub mod transform_sort_partial;
 
 pub use transform::*;
@@ -24,4 +25,5 @@ pub use transform_block_compact::*;
 pub use transform_compact::*;
 pub use transform_expression::*;
 pub use transform_expression_executor::ExpressionExecutor;
+pub use transform_sort_merge::*;
 pub use transform_sort_partial::*;
