+++ src/query/service/src/pipelines/processors/transforms/mod.rs
@@ -20,6 +20,7 @@ use common_pipeline_transforms::processors::transforms::transform;
 use common_pipeline_transforms::processors::transforms::transform_block_compact;
 use common_pipeline_transforms::processors::transforms::transform_compact;
 use common_pipeline_transforms::processors::transforms::transform_expression;
+use common_pipeline_transforms::processors::transforms::transform_sort_merge;
 use common_pipeline_transforms::processors::transforms::transform_sort_partial;
 mod transform_cast_schema;
 mod transform_create_sets;
@@ -33,7 +34,6 @@ mod transform_limit_by;
 mod transform_mark_join;
 mod transform_project;
 mod transform_rename;
-mod transform_sort_merge;
 mod transform_window_func;
 
 pub mod group_by;
